requirements.txt
aiogram>=3.17.0
aiohttp>=3.11.0
sqlalchemy>=2.0.36
asyncpg>=0.30.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
structlog>=24.4.0
python-dotenv>=1.0.1
fluent.runtime>=0.4.0
python-dateutil>=2.8.2
